Update README.md
This brings the readme to the current build of the pipeline. NOTE: Some of the dependencies listed are not necessarily need to be installed directly but I felt like it was better safe then sorry.  They need to be tested on if they are all required to be installed separately